langserv: add call and notifications access
This allows external tooling to use the JsonrpcClient when necessary.

I think we'll probably want to do as much as we can for internal libide
machanincs in IdeLangservClient rather than the plugins. Otherwise, we
have to come up with exotic caching in external structures rather than
just in the LangservClient directly.